requirements.txt
asgiref
certifi
chardet
charset-normalizer
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
idna
jmespath
loguru
psycopg
psycopg-binary
PyJWT
python-dateutil
python-dotenv
pytz
requests
s3transfer
six
sqlparse
typing_extensions
urllib3
